requirements.txt
Django>=4.0,<5.0
psycopg2-binary>=2.8
aiopokeapi>=0.1.4
